requirements.txt
langgraph
langchain
langsmith
langchain-groq
python-dotenv>=1.0.0
typing-extensions>=4.8.0
streamlit>=1.28.0
langchain-core>=0.3.0